+++ packages/typography/src/index.tsx
@@ -1,0 +1,4 @@
+export {Text, TextProps} from "./components/text"
+export {Heading, HeadingProps} from "./components/heading"
+export {Strong, StrongProps} from "./components/strong"
+export {Paragraph, ParagraphProps} from "./components/paragraph"