+++ changelog.md
@@ -1,7 +1,18 @@
-## Unreleased (2019-03-30)
+## 2019-03-31
+
+#### :rocket: New Feature
+* `typography`
+  * [#7](https://github.com/eyedea-io/smashing-ui/pull/7) Add Typography/Paragraph component ([@Idered](https://github.com/Idered))
+  * [#6](https://github.com/eyedea-io/smashing-ui/pull/6) Add Typography/Strong component ([@Idered](https://github.com/Idered))
+* `theme`, `typography`
+  * [#5](https://github.com/eyedea-io/smashing-ui/pull/5) Add Typography/Heading component ([@Idered](https://github.com/Idered))
+* `build-utils`, `theme`, `typography`
+  * [#3](https://github.com/eyedea-io/smashing-ui/pull/3) Add theme and typography packages ([@Idered](https://github.com/Idered))
 
 #### :house: Internal
-* [#2](https://github.com/eyedea-io/smashing-ui/pull/2) Add changelog generator ([@Idered](https://github.com/Idered))
+* [#4](https://github.com/eyedea-io/smashing-ui/pull/4) Improve release command ([@Idered](https://github.com/Idered))
+* [#2](https://github.com/eyedea-io/smashing-ui/pull/2) Add changelog generator
+([@Idered](https://github.com/Idered))
 * [#1](https://github.com/eyedea-io/smashing-ui/pull/1) Add Storybook ([@Idered](https://github.com/Idered))
 
 #### Committers: 1